SERIES_FLAG_UTF8_NONWORD => SERIES_FLAG_IS_SYMBOL
During the phases of transitioning to UTF-8 Everywhere, there were
incremental steps that had different flag states.  It wasn't clear
how ANY-WORD! and ANY-STRING! would be similar or different.  In
the end, string series were just constrained forms of word series.

It makes more sense for the flag to represent the constraint (as it
is added but never subtracted) and also is clearer to read.  So this
changes tracking strings as "nonwords" to saying they are symbols.