Workaround double-specification in PARSE
The way PARSE works is that if a WORD! was encountered, it would be
derelativized and what it looked up to would be used...potentially
as a rule.  The problem is that with blocks and groups this could mean
using the specifier that was intended for the array on a fetched cell.
That would make virtual binding act through a word reference, which
it's not supposed to do.

This is a hack just to get things working...which makes the fetching
of cells always put them in the same place.  Being in that place means
they're not in the array, so they've already been specified.