Formally bring back case-insensitivity of object keys
In order to make progress on virtual binding, more per-WORD! bits had
to be freed up for use.  To get enough bits to put another pointer
sized field into word cells, the spelling of bound words was derived
from the key of the object.  But since different references to the
same object field could be varied cases, that was lossy.

To avoid losing the case of words once they were bound, binding was
temporarily tried as being case sensitive.  Trying it for a time helped
to reveal what aspects of the system relied on case-insensitivity to
take a proper inventory of it--and to question how essential the
somewhat atypical feature was.

The conclusion of that process was that case-insensitivity was an
intrinsic property of binding.  This commit is a step toward bringing
that back, but with a stronger idea of canonization...so that object
keys are always the lowercase version of a symbol.  This means the
canon hash table only contains the lowercase forms, with the variants
in a linked list.

To help make things more clear, a subclass of symbol is added as a
datatype for the canon symbol forms.  This helps type check cases
where a lowercase symbol is expected.