Add passing symbol back into bound word init
In order to free up a slot in words, it was assumed that the spelling
of a word could come entirely from the binding.  This meant that when
initializing a bound word, passing in a spelling is redundant.

With the return of case insensitivity--and all keys canonized to
lowercase forms, bound words with spelling variations will need to pass
in the variation.  The plan is that this will incur some cost if there
are too many variations extant.

As a first step, this puts back passing in the variation to the bound
form of word initialization.