Change Module to a "Sea of Words"
This is an exploratory version of a change which makes it possible for
modules to "attach" bindings without actually declaring the words for
them.  This permits them to see definitions which come along later in
time, saves on variable storage space, and allows error conditions on
the assignments of things that have been misspelled.

To allow testing the concept, this introduces a linked list of patches
from a symbol to all its true definitions (not merely attached bindings)
Ultimately this would be the actual variable storage, but for now it
just contains an integer pointing to the module's context...this lets
modules still go through the ordinary enumeration process.