Resurrect CHANGE-DIR and WHAT-DIR hacks
DO has some high-level functionality of changing the working directory
of a script to be its own directory, so that FILE! paths resolve
relative to it.  This makes DO dependent on the notion of a filesystem,
which is a rough dependency to require.

As a way of working around this, the WHAT-DIR and CHANGE-DIR functions
were stubbed in so there's *some* functionality for the feature even
in a build that doesn't use the filesystem (e.g. the web build).

While hacky, it kept things running in a way that you could pull out
the notion of a filesystem and still use DO (e.g. the early web build).
Later web builds with support for schemes and web storage built on the
hack, with its own override of CHANGE-DIR.

To get Sea of Words working, the hack was removed until it was far
enough along to support it.  Now it can, so it has to be brought back
in order to test the web build.

(It is not being squashed into the history, as progressive steps in
Sea of Words require it removed, and it's desirable to still be able
to revert the intermediate states.)