Do not IMPORT a module when DO-ing it
The relationship between DO and IMPORT needs more thought...as the
DO of a module doesn't make the most sense in a world where it is
assumed modules don't have side-effects.  However, modules are right
now the means of achieving isolation in DO.  So it's a bit mixed up.

For now, being able to DO something and not know if the code is a
module or a script is used, e.g. to kick off the web build.