Make USER context at end of boot, as "normal" module
While there's been significant progress in the move away from putting
tons of things in the user context by default, it still has a couple of
assumed uses in the system.  One is for API calls in code using
libRebol that is run outside of the body of a native (e.g. code run
in `int main`).  Rather than run that in lib--which is particularly
risky--it puts it somewhere else.  The user context is also where
JS-NATIVE and TCC natives run their code, since they aren't sure
exactly where they are being run from.

But having it be built in and created through a system process makes
it "weird" in a way it shouldn't be.  e.g. it doesn't get the default
IMPORT and INTERN commands.

This switches to creating it via MODULE in the bootstrap process, so
it has all the options for creating modules available.  But this means
that code running to do the bootstrap can't use the user context...
so it uses Lib_Context during boot as a fallback.