Kill redundant module vars (WIP: can't enumerate module)
The first attempt at implementing the "sea of words" concept for
modules stored a linked list of indexes to variables declared in a
module.  Those stubs could have stored the variables themselves, but
so much of the context mechanics relied on assuming that the context
was a parallel list of keys and variables that only a small part of
the lookup was ready to use the indices.

This switches over to where the stubs actually store the data for
the variable.  This interferes with enumeration, so that's not done
yet...all modules look empty.  But it's a good first step.