Work-in-Progress: "Propagated Module Context"
This turns around the idea of how the binding process gets started from
having all the declarations being given a "baseline" binding, to
where instead everything starts out with no binding...and the very
initial "wave" of evaluation is primed with the module in which code
is running.

While it's very speculative, it boots...and so does the web console.
It also adds the ability to access parameters of a native by name in
API functions (it's important enough to be able to do it--figuring out
how to do it efficiently is worth it).